Fix multi recipe

+++ WzComparerR2.Common/CharaSim/ItemStringHelper.cs
@@ -29,7 +29,7 @@ namespace WzComparerR2.CharaSim
             }
         }
 
-        public static string GetGearPropString(GearPropType propType, int value)
+        public static string GetGearPropString(GearPropType propType, long value)
         {
             return GetGearPropString(propType, value, 0);
         }
@@ -40,7 +40,7 @@ namespace WzComparerR2.CharaSim
         /// <param Name="propType">表示装备属性枚举GearPropType。</param>
         /// <param Name="Value">表示propType属性所对应的值。</param>
         /// <returns></returns>
-        public static string GetGearPropString(GearPropType propType, int value, int signFlag)
+        public static string GetGearPropString(GearPropType propType, long value, int signFlag)
         {
 
             string sign;
@@ -657,7 +657,7 @@ namespace WzComparerR2.CharaSim
             return string.Join("､", extraJobNames) + "着用可能";
         }
 
-        public static string GetItemPropString(ItemPropType propType, int value)
+        public static string GetItemPropString(ItemPropType propType, long value)
         {
             switch (propType)
             {
@@ -1105,7 +1105,7 @@ namespace WzComparerR2.CharaSim
             return null;
         }
 
-        private static string ToCJKNumberExpr(int value)
+        private static string ToCJKNumberExpr(long value)
         {
             var sb = new StringBuilder(16);
             bool firstPart = true;
@@ -1116,14 +1116,14 @@ namespace WzComparerR2.CharaSim
             }
             if (value >= 1_0000_0000)
             {
-                int part = value / 1_0000_0000;
+                long part = value / 1_0000_0000;
                 sb.AppendFormat("{0}億", part); // Korean: 억, TradChinese+Japanese: 億, SimpChinese: 亿
                 value -= part * 1_0000_0000;
                 firstPart = false;
             }
             if (value >= 1_0000)
             {
-                int part = value / 1_0000;
+                long part = value / 1_0000;
                 sb.Append(firstPart ? null : " ");
                 sb.AppendFormat("{0}万", part); // Korean: 만, TradChinese: 萬, SimpChinese+Japanese: 万
                 value -= part * 1_0000;
